+++ ElectronicCard/EC.BusinessLogic/Services/Implementation/PreparationService.cs
@@ -16,22 +16,37 @@ namespace EC.BusinessLogic.Services.Implementation
 
         public void CreatePreparation(Preparation preparation)
         {
+            if (preparation == null)
+            {
+                return;
+            }
+
             _preparationRepository.Create(preparation);
         }
 
         public void UpdatePreparation(Preparation preparation)
         {
+            if (preparation == null)
+            {
+                return;
+            }
+
             _preparationRepository.Update(preparation);
         }
 
         public void DeletePreparation(int? id)
         {
+            if (id == null || id <= 0)
+            {
+                return;
+            }
+
             _preparationRepository.Delete(id);
         }
 
         public Preparation GetById(int? id)
         {
-            return id == null ? null : _preparationRepository.GetById(id);
+            return id == null || id <= 0 ? null : _preparationRepository.GetById(id);
         }
 
         public IReadOnlyCollection<Preparation> GetAll()
